pytest: move wallet tests into new file tests/test_wallet.py
We're going to add some more.
